# openapi.yaml
openapi: 3.0.3
info:
  title: OpenCode Manager API
  description: |
    Global OpenCode server manager for multi-project workflows.
    
    This API manages multiple OpenCode AI assistant servers across different projects, allowing you to:
    - Start/stop OpenCode servers per project
    - Manage tmux TUI (Terminal User Interface) attachments
    - Send prompts to project-specific OpenCode instances
    - Stream real-time events via Server-Sent Events (SSE)
    
    **Important:** Project paths in URLs must be base64url-encoded.
  version: 1.0.0
  contact:
    name: OpenCode Manager
servers:
  - url: http://127.0.0.1:4095
    description: Local development server

tags:
  - name: Health
    description: Health check and status endpoints
  - name: MCP
    description: Model Context Protocol endpoint (read-only documentation for API discovery)
  - name: Projects
    description: Project registry and status management
  - name: Server Management
    description: Start, stop, and manage OpenCode servers
  - name: Prompts
    description: Send prompts to OpenCode instances
  - name: TUI
    description: Terminal UI attachment and management
  - name: Events
    description: Server-Sent Events streaming

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns health status and version information
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  healthy:
                    type: boolean
                    example: true
                  version:
                    type: string
                    example: "1.0.0"

  /mcp:
    post:
      tags:
        - MCP
      summary: MCP Documentation Endpoint - POST
      description: |
        **Read-Only Documentation Endpoint** implementing MCP Streamable HTTP (2025-06-18).
        
        Exposes REST API documentation as MCP Resources. **NO executable actions.**
        AI agents discover the API structure, then use REST endpoints for operations.
        
        **Required Headers:**
        - `MCP-Protocol-Version`: Protocol version (e.g., "2025-06-18")
        - `Accept`: Must include "application/json" and/or "text/event-stream"
        - `Mcp-Session-Id`: Session ID (required after initialization)
        
        **Supported Methods:**
        - `initialize` - Initialize session (returns Mcp-Session-Id header)
        - `resources/list` - List available documentation resources
        - `resources/read` - Read documentation resource content
        - `notifications/initialized` - Acknowledge initialization (returns 202)
        - `notifications/cancelled` - Cancel request (returns 202)
        
        **Available Resources:**
        - `resource://opencode-manager/openapi` - Full OpenAPI spec (YAML)
        - `resource://opencode-manager/api/endpoints` - Endpoint summary (JSON)
        - `resource://opencode-manager/api/examples` - Usage examples (Markdown)
        - `resource://opencode-manager/architecture` - System architecture (Markdown)
        
        **Response Types:**
        - Single JSON response (Content-Type: application/json)
        - SSE stream (Content-Type: text/event-stream) - based on Accept header preference
        - 202 Accepted for notifications
      operationId: mcpPost
      parameters:
        - name: MCP-Protocol-Version
          in: header
          required: true
          schema:
            type: string
            enum: ["2025-06-18", "2025-03-26", "2024-11-05"]
          description: MCP protocol version
          example: "2025-06-18"
        - name: Accept
          in: header
          required: true
          schema:
            type: string
          description: Must include application/json and/or text/event-stream
          example: "application/json, text/event-stream"
        - name: Mcp-Session-Id
          in: header
          required: false
          schema:
            type: string
          description: Session ID (required after initialization)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - jsonrpc
                - method
              properties:
                jsonrpc:
                  type: string
                  enum: ["2.0"]
                method:
                  type: string
                params:
                  type: object
                id:
                  oneOf:
                    - type: string
                    - type: number
            examples:
              initialize:
                summary: Initialize session
                value:
                  jsonrpc: "2.0"
                  method: "initialize"
                  params:
                    protocolVersion: "2025-06-18"
                    clientInfo:
                      name: "example-client"
                      version: "1.0.0"
                  id: 1
              listResources:
                summary: List documentation resources
                value:
                  jsonrpc: "2.0"
                  method: "resources/list"
                  params: {}
                  id: 2
              readResource:
                summary: Read a documentation resource
                value:
                  jsonrpc: "2.0"
                  method: "resources/read"
                  params:
                    uri: "resource://opencode-manager/api/endpoints"
                  id: 3
      responses:
        '200':
          description: JSON-RPC response or SSE stream
          headers:
            Mcp-Session-Id:
              description: Session ID (returned on initialize)
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  jsonrpc:
                    type: string
                    example: "2.0"
                  id:
                    oneOf:
                      - type: string
                      - type: number
                  result:
                    type: object
            text/event-stream:
              schema:
                type: string
              example: |
                id: session-123-0
                data: {"jsonrpc":"2.0","id":1,"result":{...}}
        '202':
          description: Accepted (for notifications and responses)
        '400':
          description: Bad request (invalid protocol version, missing headers, or invalid JSON-RPC)
        '404':
          description: Session not found or expired

    get:
      tags:
        - MCP
      summary: MCP Streamable HTTP - GET endpoint
      description: |
        Server-Sent Events stream for server-initiated messages (MCP 2025-06-18).
        
        Opens a persistent SSE connection for receiving server requests and notifications.
        
        **Required Headers:**
        - `MCP-Protocol-Version`: Protocol version
        - `Accept`: Must include "text/event-stream"
        - `Mcp-Session-Id`: Session ID (required)
        - `Last-Event-Id`: Optional, for resuming broken connections
        
        **Features:**
        - Resumable streams with event IDs
        - Automatic keepalive pings
        - Session-based message delivery
      operationId: mcpGet
      parameters:
        - name: MCP-Protocol-Version
          in: header
          required: true
          schema:
            type: string
          example: "2025-06-18"
        - name: Accept
          in: header
          required: true
          schema:
            type: string
          example: "text/event-stream"
        - name: Mcp-Session-Id
          in: header
          required: true
          schema:
            type: string
          description: Session ID from initialization
        - name: Last-Event-Id
          in: header
          required: false
          schema:
            type: string
          description: Last received event ID for resuming stream
      responses:
        '200':
          description: SSE stream established
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                id: session-123-5
                : connection established
                
                : ping
        '400':
          description: Invalid protocol version or missing Accept header
        '404':
          description: Session not found or expired
        '405':
          description: Method not allowed (Accept header must include text/event-stream)

    delete:
      tags:
        - MCP
      summary: MCP Session termination
      description: |
        Explicitly terminate an MCP session.
        
        Closes all open SSE streams and removes session data.
        
        **Required Headers:**
        - `Mcp-Session-Id`: Session ID to terminate
      operationId: mcpDelete
      parameters:
        - name: Mcp-Session-Id
          in: header
          required: true
          schema:
            type: string
          description: Session ID to terminate
      responses:
        '200':
          description: Session terminated successfully
        '400':
          description: Missing Mcp-Session-Id header
        '404':
          description: Session not found

  /projects:
    get:
      tags:
        - Projects
      summary: List all projects
      description: Returns the complete project registry with all registered projects and their status
      operationId: listProjects
      responses:
        '200':
          description: Project registry
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/ProjectEntry'

  /project/{path}:
    get:
      tags:
        - Projects
      summary: Get project status
      description: Returns detailed status information for a specific project
      operationId: getProjectStatus
      parameters:
        - $ref: '#/components/parameters/ProjectPath'
      responses:
        '200':
          description: Project entry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectEntry'
        '400':
          $ref: '#/components/responses/InvalidPathEncoding'
        '404':
          $ref: '#/components/responses/ProjectNotFound'
    
    delete:
      tags:
        - Server Management
      summary: Stop OpenCode server
      description: Stops the OpenCode server for the specified project
      operationId: stopServer
      parameters:
        - $ref: '#/components/parameters/ProjectPath'
      responses:
        '200':
          description: Server stopped successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/InvalidPathEncoding'
        '500':
          $ref: '#/components/responses/ServerError'

  /project/{path}/ensure:
    post:
      tags:
        - Server Management
      summary: Ensure server is running
      description: Starts the OpenCode server if not running, or returns the existing port if already running
      operationId: ensureServer
      parameters:
        - $ref: '#/components/parameters/ProjectPath'
      responses:
        '200':
          description: Server is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  port:
                    type: integer
                    example: 4097
                  status:
                    type: string
                    enum: [running]
                    example: running
        '400':
          $ref: '#/components/responses/InvalidPathEncoding'
        '500':
          $ref: '#/components/responses/ServerError'

  /project/{path}/prompt:
    post:
      tags:
        - Prompts
      summary: Send prompt to project
      description: |
        Sends a prompt to the project's OpenCode server. The server will be started automatically if not running.
        
        Returns 204 No Content immediately. Listen to `/project/{path}/events` for real-time updates and responses.
      operationId: sendPrompt
      parameters:
        - $ref: '#/components/parameters/ProjectPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: The prompt text to send
                  example: "Add a new feature to handle user authentication"
                file:
                  type: string
                  description: Optional file path for context
                  example: "/path/to/project/src/auth.ts"
                line:
                  type: integer
                  description: Optional line number in the file
                  example: 42
                col:
                  type: integer
                  description: Optional column number in the file
                  example: 10
      responses:
        '204':
          description: Prompt sent successfully (listen to events endpoint for updates)
        '400':
          description: Bad request (missing text or invalid path)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /project/{path}/attach-tui-cli:
    post:
      tags:
        - TUI
      summary: Attach TUI in current pane (CLI)
      description: |
        Attaches the OpenCode TUI in the current tmux pane. This is intended for CLI usage where the TUI should replace the current shell.
        
        Requires tmux session named 'dev' to be running.
      operationId: attachTuiCli
      parameters:
        - $ref: '#/components/parameters/ProjectPath'
      responses:
        '200':
          description: TUI attached successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  pane:
                    $ref: '#/components/schemas/TmuxPane'
        '400':
          description: Tmux session not running or invalid path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /project/{path}/attach-tui-neovim:
    post:
      tags:
        - TUI
      summary: Attach TUI from Neovim
      description: |
        Smart TUI attachment from Neovim:
        - If TUI already exists for this project, focuses the existing pane
        - Otherwise, creates a new pane and attaches the TUI
        
        Requires tmux session named 'dev' to be running.
      operationId: attachTuiNeovim
      parameters:
        - $ref: '#/components/parameters/ProjectPath'
      responses:
        '200':
          description: TUI focused or created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  focused:
                    type: boolean
                    description: True if existing pane was focused
                  created:
                    type: boolean
                    description: True if new pane was created
                  pane:
                    $ref: '#/components/schemas/TmuxPane'
        '400':
          description: Tmux session not running or invalid path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: TUI pane no longer exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /project/{path}/focus-tui:
    post:
      tags:
        - TUI
      summary: Focus existing TUI pane
      description: Focuses an existing TUI pane for the project
      operationId: focusTui
      parameters:
        - $ref: '#/components/parameters/ProjectPath'
      responses:
        '200':
          description: TUI focused successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '400':
          $ref: '#/components/responses/InvalidPathEncoding'
        '404':
          description: No TUI attached for this project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /project/{path}/events:
    get:
      tags:
        - Events
      summary: Server-Sent Events stream
      description: |
        Subscribe to real-time events from the OpenCode server for a specific project.
        
        Event types:
        - `connected`: Initial connection established
        - `message.created`: New message created
        - `message.updated`: Message content updated
        - `message.completed`: Message fully completed
        - `part.updated`: Message part updated
        - `permission.requested`: Permission request from OpenCode
        - `session.updated`: Session state changed
      operationId: getEvents
      parameters:
        - $ref: '#/components/parameters/ProjectPath'
      responses:
        '200':
          description: SSE stream established
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream
              example: |
                event: connected
                data: {"projectPath":"/path/to/project"}
                
                event: message.created
                data: {"id":"msg_123","text":"Hello"}
        '400':
          $ref: '#/components/responses/InvalidPathEncoding'
        '404':
          description: No running server for this project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  parameters:
    ProjectPath:
      name: path
      in: path
      required: true
      description: Base64url-encoded project path
      schema:
        type: string
      example: L1VzZXJzL2V4YW1wbGUvcHJvamVjdA

  schemas:
    ProjectEntry:
      type: object
      properties:
        port:
          type: integer
          description: Port number where the OpenCode server is running
          example: 4097
        pid:
          type: integer
          nullable: true
          description: Process ID of the OpenCode server
          example: 12345
        status:
          type: string
          enum: [starting, running, stopped]
          description: Current server status
          example: running
        hasTUI:
          type: boolean
          description: Whether a TUI is currently attached
          example: true
        tmuxPane:
          type: string
          nullable: true
          description: Tmux pane identifier (format - session:window.pane)
          example: "dev:projectname.1"
        createdAt:
          type: string
          format: date-time
          description: When the project was first registered
        lastUsed:
          type: string
          format: date-time
          description: Last time the project was accessed

    TmuxPane:
      type: object
      properties:
        session:
          type: string
          description: Tmux session name
          example: dev
        window:
          type: string
          description: Tmux window name
          example: projectname
        paneId:
          type: string
          description: Tmux pane ID
          example: "%1"
        paneIndex:
          type: integer
          description: Pane index within the window
          example: 1

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Project not found"

  responses:
    ProjectNotFound:
      description: Project not found in registry
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Project not found"

    InvalidPathEncoding:
      description: Invalid base64url path encoding
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid path encoding"

    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"